Refactoring: renamed importers to exporters
We originally used the term "importer" to refer to components that
wrote graphs to the database, that is, these components import into
the database. However, since we now support writing of these
structures to disk in text formats, it is much more natural to think
of these components as "exporters".