CSV keys such as type:flags[] or funcid:int are not only identified by their name, but by their name as well as their type.
Sometimes this is important. In particular, the header format used by the edge files consists of type specifications rather than names. It is

    :START_ID,:END_ID,:TYPE

and *not*

    START_ID,END_ID,TYPE

See http://neo4j.com/docs/stable/import-tool-header-format.html as well as http://neo4j.com/docs/stable/import-tool-examples.html

The previous architecture did not allow us to import such unnamed-but-typed fields from CSV files.

Therefore I had to change the class KeyedCSVRow to maintain a HashMap<CSVKey,String> of values instead of a HashMap<String,String>. This feels "cleaner" anyway. :-) After all, the CSVKey is the key, as the name of the class itself says...

For this to work properly, the CSVKey class has to override the standard equals(Object) and hashCode() methods. Some tests have additionally been added in a new class TestCSVKey. Some other classes had to be changed accordingly as well.

The TestCSV2AST class is also of interest here, since it is now able to use the proper header format, and in particular testSimpleEdgeImport() works.

All of this will allow us to parse edge CSV files exactly as output by the phpjoern parser. In particular, the foundation is laid for the CSVFunctionExtractor to be able to import edges as well (note that it currently only imports nodes, but never connects them, as it never even reads from the edges CSV file, except for the header line). I'm working on this now; once this is done, the CSVFunctionExtractor will be finished.