Added tests for function-declaring nodes:
* AST_TOPLEVEL -> ast.php.functionDef.TopLevelFunctionDef
* AST_FUNC_DECL -> ast.functionDef.FunctionDef
* AST_CLOSURE -> ast.php.functionDef.Closure
* AST_METHOD -> ast.php.functionDef.Method

Additionally, extended FunctionDef with a new method getReturnType() that returns the Identifier child representing the function's return type.
It came in handy to use the already defined Identifier child, since I indeed do need an Identifier for function nodes for PHP ASTs,
but for the function's return type and not for the function's name... I could have declared another Identifier child called returnType instead
(in fact, this was my first approach), but then it gets tricky when it comes to the method addChild(ASTNode) in FunctionDef.
Indeed, in this case it would need to somehow check whether the Identifier child that's being added is the name of the method
(C ASTs) or the name of the return type (PHP ASTs). Since each AST needs exactly one of the two Identifiers, that would be
unnecessarily complicated.