Implemented smart edge interpreter.
Before, the logic for adding childs was implemented entirely in the different kinds of nodes (classes extending ASTNode) itself (e.g., in PHPParameter: "Am I adding a parameter type? Or a parameter name? Or a parameter default value?"), within the addChild() method. This will work fine provided that for each kind of node, it holds that the node's children all have pairwise distinct node types. Then, the addChild() method can always call the appropriate handler by looking at the child's type.

For PHP ASTs, this is not always given. There are cases where a particular kind of node can have different children of the same kind. An example of this is the AST_PARAM node with its 'name' and 'default' children. When this occurs, the only way that the `addChild()` method can *halfway* reliably distinguish what handler to call is by the order in which these children were added. But this assumes that the addChild() will be called a certain number of times with different parameters *in a particular, undocumented order* (it requires that the edges in the edges file appear in a certain order.) This feels like a hack, and it's error-prone.

Thinking about this, I realized that this logic belongs into the edge interpreter. The edge interpreter knows the AST, it knows the start node id and the end node id (and, thus, the start node and the end node itself), it can look at the childnum property, and it knows exactly what kind of child it is adding. Once this dawned on me, it became plain as day: The node interpreter knows what kind of nodes it is adding, and the edge interpreter knows what kind of edges it is adding. Thus I implemented a smart edge interpreter which does exactly that, and directly calls a node's appropriate handlers instead of generically calling the addChild() method. Additionally, thereby the PHP AST structure is (/will) now be documented in these two central files: the PHPCSVNodeInterpreter and the PHPCSVEdgeInterpreter. The only case where the addChild() method still makes sense for PHP ASTs is in those kind of nodes which have an arbitrary number of children, such as CompoundStatement or ParameterList.

Conversely, a particular child of a PHP AST node that can have different kinds. For instance, a "return type" child of a function node can be either an Identifier node if the corresponding PHP function returns something, or a "null" node if it is a void function. Such a null node, i.e., an ASTNode with its type set to NULL, should be added to the function node's children and count towards the children count, but it is not an Identifier. Hence in such a case, FunctionDef.getReturnType() can only return the null value instead of the null node itself, which is fine. The point is that the previous architecture behaved in the same way, but did so in a kind of implicit, somewhat hidden way; the new architecture makes this reasoning explicit.

Note that though technically possible, it would not be a good idea to have the logic that calls handlers according to their childnum property in the addChild() methods of the different ASTNode's. The childnum property is not an inherent property of a PHP AST node: Rather, like funcid, it is a meta-property pertaining to the CSV files. Therefore, even though the childnum property is currently read from CSV and saved to any new ASTNode by the node interpreter, the addChild() method of an ASTNode should not "know" about it. In fact, the childnum property rather belongs in the edges file, not in the nodes file, and in time I will modify the phpjoern exporter to do just that. Then, the childnum property will no longer be added to the ASTNode's created by the node interpreter (just like the funcid isn't, either). Additionally, it would get very ugly if we tried to consolidate an ASTNode's addChild() method for both C and PHP ASTs by having it look first whether the childnum property is defined (-> we are in PHP world) and call the appropriate handlers according to that, or if it is not defined (-> we are in C world), look at the child's type and call the appropriate handlers according to that, or something like that.

It was worth doing this, since it will ease work a lot in the future. Before it was always kind of a headache trying to consolidate the addChild() method for both C and PHP ASTs, which use different children for different purposes. A prominent recent example is the Identifier child of a Parameter, which represents the parameter's name in C world and the parameter's type in PHP world. Another such example is a FunctionDef's Identifier child, which represents the function's name in C world and the function's return type in PHP world. Consolidating the addChild() method for both worlds will now no longer be necessary. I can leave the addChild() method more or less untouched, as I do not need in any more. I am saying "more or less" because I do need the addChild() method of the super-class ASTNode that adds any child to a node's linked list of children, but this super-method is now always directly called from the appropriate handlers; in order to avoid double addition, the addChild() methods of the different classes extending ASTNode now only call the super-method addChild() if an appropriate handler method has not been called, since the handler method already called the super-method addChild(). Hence, for the C world everything remains as before, but it is now also possible to call the handler methods directly, and that is what the PHP edge interpreter is doing.

Putting the logic for distinguishing children in the addChild() methods might be good for C ASTs, what with the node visitor and tree walking and all that (I assume -- I'm not all that familiar with that part of Joern), but for PHP, where we read the data out of CSV files, it is not the right approach.