Added support for AST_NAME_LIST nodes.
* AST_NAME_LIST -> ast.expressions.IdentifierList

I put IdentifierList in the ast.expressions package (instead of the ast.php.expressions package) because I felt that the concept is general enough to be useful for other langages as well. For instance, if we want to implement support for Java sometime, then IdentifierList's might be equally useful for storing a list of interfaces implemented by a class (as is the case for PHP).