Completed support for AST_WHILE
* AST_WHILE -> ast.statements.blockstarters.WhileStatement

Note: I had to change the type of BlockStarter.condition from Condition to a generic ASTNode for now (and adapt getCondition() and setCondition() accordingly).
This is because there are so many PHP AST node types that can occur in a while loop's guard that are not (yet) mapped for now that we cannot at the moment cast these nodes to Condition, lest we get ClassCastException's.

This implies that I had to change some C test cases to downcast the ASTNode returned by getCondition() to a Condition, similarly as I had to do earlier for Parameter whose getType() method I had to change to return an ASTNode instead of a ParameterType.

In time (once all the mapping is finished), we should be able to make the getCondition() and setCondition() methods more "precise" again. I'm not sure whether we will want them to return Condition's again or rather Expression's; but we will see about that then.

Note #2: Conversely, I introduced getContent() and setContent() methods in ast.statements.blockstarters.BlockStarter that do pretty much the same thing as getStatement() and setStatement(), but are *more* precise than getStatement() and setStatement() in that they take resp. return a CompoundStatement (which extends Statement). Indeed, for PHP ASTs, we know that the second child of an AST_WHILE node is always going to be an AST_STMT_LIST.