Added a new class TestPHPCSVASTBuilderMinimal
In view of some recent developments, I felt it was a good idea to introduce a new test class
that generates some AST nodes that have as few children as possible (as opposed to the "normal"
test class TestPHPCSVASTBuilder which aims to generate all possible children of any AST node),
so we don't get NullPointerExceptions flying around our ears once we throw Joern on real-world
PHP code.

For instance, while and do-while statements may have a "statement" child that is only a single
statement and thus is not an AST_STMT_LIST. This can even be a no-op statement, composed
only of a semicolon, which generates a null node.

Another interesting insight was that methods can have no AST_STMT_LIST body at all (e.g., method
signatures defined in an abstract class.); this does not hold for functions or closures (these
always have an AST_STMT_LIST body.)

For the constructs generated so far, we now support this properly.