Added support for return statements.
* AST_RETURN -> ast.statements.jump.ReturnStatement

I extended the ReturnStatement class with a returnExpression field, as well as corresponding getter- and setter-methods.

In this case I chose to change the class ReturnStatement instead of having a PHPReturnStatement extend it, as return expressions
also occur in C and other languages, so it makes sense to have this in the base class. I am unsure how a return statement's expression
is currently handled in Joern C world. I assume using generic addChild() and getChild() calls... Maybe the C world could be refactored
to use the new getter- and setter-methods, but this is rather cosmetic really.