Added support for goto and label statements.
* AST_LABEL -> ast.logical.statements.Label
* AST_GOTO -> ast.statements.jump.GotoStatement

Similarly as for ReturnStatement, I enriched Label with getNameChild() and setNameChild() methods, as well
as GotoStatement with getTargetLabel() and setTargetLabel() methods. They are generic enough to be used
by other languages, even though they are not used in C world.

I also renamed the already present method getTarget() to getTargetName() for consistency.