Added support for catch statements.
* AST_CATCH -> ast.statements.blockstarters.CatchStatement
* AST_CATCH_LIST -> ast.statements.blockstarters.CatchList

Note that CatchList is new. The existence of this node type is consistent with similar constructs such as ParameterList or SwitchList, and required for PHP ASTs. I added it to the general ast.statements.blockstarters package (instead of ast.php.statements.blockstarters) since the concept is general enough to be useful for other languages.