Added support for try statements.
* AST_TRY -> ast.statements.blockstarters.TryStatement

Note that I added a field finallyContent to represent a finally clause to the base class TryStatement. Even though this is irrelevant for C, it is, once again, general enough to be useful for other languages than PHP (e.g., Java, Python, ...), so I felt it would be weird to have a class PHPTryStatement that extends TryStatement only to add support for a finally clause. Anyhow, I was a bit astonished when I saw that there was already a class TryStatement in the first place. I thought that C does not have try/catch constructs at all?

Additionally, note that we definitely have to do some refactoring here. The problem is that PHP ASTs represent try statements as a list with three nodes: the content to be tried; a list of catch statements; and a finally clause.
For C ASTs however, it appears that catch statements are managed directly as children of a try statement node. This means that in C ASTs, try statements can have an arbitrary number of children, with the first being the content to be tried, and all following children being catch nodes. Basically, the CatchList abstraction present in PHP ASTs inbetween a try statement and its catch nodes is missing in C ASTs. This leads to some unclean behavior, e.g., TryStatement currently overrides getChildCount() to return "super.getChildCount() + getCatchNodes().size()", and getChild(int) relays to "getCatchNode(i - 1)" unless i is equal to 0. These overrides are unwished for in PHP world. If C world really needs them, it should use an extending class CTryStatement, but the cleanest solution would be to also use a CatchList abstraction layer in-between as is the case in PHP. Another problem arising from this way of managing children of a try statement is that it does not generalize well to include support for finally clauses.