Refactored C world to use a catch list abstraction
C world was refactored to implicitly use a CatchList of CatchStatement's instead of having a TryStatement manage its CatchStatement children directly.
I say "implicit" because the "outer world" in C does not need to instantiate a CatchList, instead a CatchList is implicitly created whenever a TryStatement is created. So basically this modification only affects how catch statement children are managed *internally* by a TryStatement in C world. The only difference is that getChildCount() should now normally be 2, instead of arbitrary: a statement and a catch list. However, all tests are still passing, I assume the number of children of a TryStatement was never tested in C world.

Additionally, I also changed the line
  if (tryStatement.getCatchNodes() == null)
to be
  if (tryStatement.getCatchList().size() == 0)
in CCFGFactory, since I assume that's what was meant. Indeed, previously TryStatement always declared
  private List<CatchStatement> catchNodes = new LinkedList<CatchStatement>();
so tryStatement.getCatchNodes() would *never* be equal to null and therefore the code after the if statement was unreachable.