Added support for call expressions.
* AST_CALL -> ast.expressions.CallExpression

I added the fields target and argumentList to CallExpression, with corresponding getters and setters.
I also refactored the method getTarget() to return this.target instead of getChild(0). Thus I had to override the method addChild(ASTNode) even though PHP does not need it, so that the getTarget() call continues working in C world.