Added support for static call expressions.
* AST_STATIC_CALL -> ast.php.expressions.StaticCallExpression

Had to use a super-ugly hack in ast.expressions.CallExpression now because the addChild(ASTNode) is making problems again.
We only need this method override for C world: it is supposed to call the setTarget(ASTNode) method to set the target function's name.
In PHP world, setTarget(ASTNode) is called directly anyway.
However, StaticCallExpression#setTargetClass(Identifier) needs to call super.addChild(ASTNode) in order to add the target
class's name to the list of children of the StaticCallExpression. It doesn't mean to call CallExpression#addChild(ASTNode) which
is only there for C; rather, it means to call ASTNode#addChild(ASTNode). But because it inherits from CallExpression it cannot avoid
this detour, and this detour then calls CallExpression#setTarget(ASTNode) with the class name identifier, because it notices that the child being
added is an Identifer... thus, I had to introduce the additional condition this.getClass().equals(CallExpression.class) in
CallExpression#addChild(ASTNode), only to make sure we are not in a sub-class, which is super-ugly!

TODO: refactor C world to use a CCallExpression which overrides addChild(ASTNode) on its own. That's the clean way.