Added support for constant declarations.
* AST_CONST_ELEM -> ast.php.statements.ConstantElement
* AST_CONST_DECL -> ast.php.statements.ConstantDeclaration
* AST_CLASS_CONST_DECL -> ast.php.statements.ClassConstantDeclaration

Similarly to the previous commit concerning properties, these might be useful for other OO languages, but I put them into ast.php.statements for now.
Note that constants in PHP may be declared on top level as well as within a class's scope, whereas properties are always in a  class scope.