Added support for traits.
* AST_USE_TRAIT -> ast.php.statements.blockstarters.PHPUseTrait
* AST_TRAIT_ADAPTATIONS -> ast.php.statements.blockstarters.PHPTraitAdaptations
* AST_TRAIT_ALIAS -> ast.php.statements.blockstarters.PHPTraitAlias
* AST_TRAIT_PRECEDENCE -> ast.php.statements.blockstarters.PHPTraitPrecedence
* AST_METHOD_REFERENCE -> ast.php.statements.blockstarters.MethodReference

Additionally, created a new class ast.php.statements.blockstarters.PHPTraitAdaptationElement as a common ancestor for PHPTraitAlias and PHPTraitAdaptations.

For an explanation of traits in PHP, see http://php.net/manual/en/language.oop5.traits.php

A PHPUseTrait node has an IdentifierList and (optionally) a PHPTraitAdaptations child (or null child);
a PHPTraitAdaptations node has an arbitrary number of PHPTraitAdaptationElement's as children;
a PHPTraitAlias has a MethodReference and a string child;
a PHPTraitPrecedence has a MethodReference and an IdentifierList child;
a MethodReference has an Identifier and a string child.