Added support for binary operation expressions.
* AST_BINARY_OP -> ast.expressions.BinaryOperationExpression

I introduced a new class BinaryOperationExpression which extends BinaryExpression, and made this new class a common ancestor for the following nodes:
- AdditiveExpression
- AndExpression
- BitAndExpression
- EqualityExpression
- ExclusiveOrExpression
- InclusiveOrExpression
- MultiplicativeExpression
- OrExpression
- RelationalExpression
- ShiftExpression

Thereby, the class BinaryExpression now has only two *direct* children: AssignmentExpression and BinaryOperationExpression, thus giving us a more fine-grained understanding of the various kinds of nodes that exist by making a clear distinction between assignment expressions and binary operation expressions, both of which are binary expressions.

This change was also necessary because PHP ASTs do *not* use a distinguished kind of node for each kind of binary expression, as Joern does in C world; rather, (almost) all binary expressions are represented as an AST_BINARY_OP node. The concrete operation that the node represents is specified via a flag on this node. Also see https://github.com/nikic/php-ast#flags.
Generally I think that's a good thing, because there are even more kinds of binary operation expressions in PHP than there are in C, and having a distinguished node for each of them would generate a lot of overhead. This "flag" approach is, by the way, also used for AST_ASSIGN_OP nodes (added in the previous commit).

The following is a list of PHP's binary operators that will map to an AST_BINARY_OP node. This list makes every claim to completeness:

// bit operators
$or1 | $or2;
$and1 & $and2;
$msg ^ $otp;
$x << $y;
$x >> $y;
// string operators
$str1 . $str2;
// arithmetic operators
$x + $y;
$x - $y;
$x * $y;
$x / $y;
$x % $y;
$x ** $y;
// boolean operators
$x xor $y;
// comparison operators
$x === $y;
$x !== $y;
$x == $y;
$x != $y;
$x < $y;
$x <= $y;
$x <=> $y;