Added support for greater, greater or equal, and and or expressions.
* AST_GREATER -> ast.expressions.GreaterExpression
* AST_GREATER_EQUAL -> ast.expressions.GreaterOrEqualExpression
* AST_AND -> ast.expressions.AndExpression
* AST_OR -> ast.expressions.OrExpression

I had to add GreaterExpression and GreaterOrEqualExpression as new classes inheriting from RelationalExpression.

These four nodes are extremely similar to the BinaryOperationExpression's added in the last commit.
In fact, in a future version of the php-ast extension by Niki (version 20), these nodes will not be necessary any longer, as the corresponding expressions will be mapped to BinaryOperationExpression too. The fact that they currently aren't is most probably due to historic reasons, and illustrates that PHP 7 is still under development.