Added support for echo statements.
* AST_ECHO -> ast.php.statements.PHPEchoStatement

Though print statements exist in probably pretty much every language out there, the fact that "echo" in PHP is a language construct and not a function call to some standard library makes it pretty PHP-specific.

This also finished all statements with a single child. Now: only some expressions with a single child to go, and we'll only be left with some leaf nodes to take care of. :)