Added support for type casts.
* AST_CAST -> ast.expressions.CastExpression

I refactored CastExpression to include setters for the cast target and cast expression (setCastTarget() and setCastExpression() respectively), and had them call super.addChild() respectively, as is being done in all other nodes too. Then I modified addChild(ASTNode) to call these setters. Since the children nodes are thereby *actually* being added to the node's list of children, overriding getChild(int) and getChildCount() became unnecessary, and I therefore removed these overrides.

This situation was similar to that of the BinaryExpression node in commit 062b773c64ab515510097aead0aedbc5ad70793a.

Thus the design of this class is now consistent with that of other classes, with the children being actually added to the node's list of children and without overrides for getChild(int) and getChildCount().