Added support for type hints.
* AST_TYPE -> ast.php.expressions.PHPTypeHint

This is used for the type hints 'array' and 'callable' for function parameters or return types.
It is yet unclear to me why AST_TYPE is used only for 'array' and 'callable', while other type hints such as 'int', 'double', 'string' etc. do not produce an AST_TYPE node, but an AST_NAME node. Would it not be more consistent to map all primary type hints to AST_TYPE, and use AST_NAME only for "self-defined" types, i.e., class names?