Massive code cleanup.
Killed tons of small TODOs scattered everywhere. :)

**Children now use more precise child types**

Now that the mapping is finished, we can make the fields corresponding to a node's children much more precise (and accordingly, the getter and setter methods as well). Instead of a generic ASTNode, these fields can now have specific other node types (e.g., Expression).

This was not possible before the mapping was finished, often simply because a given node type that appeared as a child of another node did not exist yet; or, e.g., because some node had a child of which we know that is always of type Expression, but since we had not yet mapped all nodes, not all nodes could be cast to Expression. To avoid ClassCastException's, we therefore had to use a generic ASTNode as the child type. At times, when I mapped a PHP AST node to a Joern AST node that already existed, I even had to make the already-existing fields "less" precise, e.g., change them from an Expression to a generic ASTNode.

Now, all nodes have fields (and getters and setters for them) that are as precise as possible, as it should be. The edge interpreter casts the children nodes to the appropriate types. This was a big TODO that I had to push off until after the mapping was finished, and it is now done. :)

**The null node**

I also introduced a new dummy "ast.NullNode":

* NULL -> ast.NullNode

Thereby, we can better deal with null nodes in PHP ASTs. Indeed, like primary expressions, there are nodes in PHP ASTs that are literally not instances of ast\Node nor plain values such as integer, string or double, but simply null!
Having a dedicated NullNode makes sense to keep the number of children of nodes with a fixed number of children constant (this is the same reason why null nodes exist in PHP ASTs). For instance, when a function node has no return type specified in its declaration, then it will still have a "return type" child which will be a NullNode. Moreover, we often have to distinguish the cases where we are dealing with a null node or with a non-null node: e.g., we can only cast a non-null child to its appropriate type (as explained above) if it is *not* a null node, and we therefore have to ensure that this child is *not* an instance of NullNode.