Created a new class PHPFunctionDef which extends FunctionDef.
As discussed in issue #80.

All other PHP function nodes (Closure, Method and TopLevelFunctionDef) now extend PHPFunctionDef instead of FunctionDef.

All code dealing with PHP function now uses PHPFunctionDef instead of FunctionDef.

In general, when dealing with PHP functions in Joern, one should now *always* work with a PHPFunctionDef instead of a FunctionDef, even when doing very "general" work and the type of funcion is not important. In particular,
- note that the *name* of a PHP function is retrieved via getName() and not getIdentifier(), and getName() is only defined by PHPFunctionDef;
- the return type of a PHP function declaration can be retrieved via getReturnType(), which is only defined by PHPFunctionDef;
- finally, getDocComment() is also only defined by PHPFunctionDef.