Added support for expression statements.
When adding an expression as a child to represent the statement executed in the body of a while/do/if/for/foreach statement,
an ExpressionStatement is automatically created which holds the given expression, and this expression statement is
added as a child for the while/do/if/for/foreach statement.

This is necessary since an expression in itself is not a statement, we need to "convert" it to an expression statement first.