Redesigned inputModules.csv.csvFuncExtractor.CSVAST to maintain two LinkedList<KeyedCSVRow> instead of two LinkedList<String>.
Earlier, when parsing a CSV file with many files and functions, things were such that
(1) the CSVFunctionExtractor would use a KeyedCSVReader to transform each CSV row of the CSV node file it was reading into a KeyedCSVRow object;
(2) for each such KeyedCSVRow, it would determine which function it belongs to, create a new CSVAST object for this function (if it did not exist yet), and add the node row to this CSVAST *as a string*, that is, transforming the readily parsed KeyedCSVRow *back* to a string!
(3) Later, the CSV2AST converter would *again* initiate a KeyedCSVReader, ask the CSVAST object to join all its node rows (which were now strings again) with "\n" so as to return a single string, then *again* parse this string anew with the KeyedCSVReader, creating KeyedCSVRow's to pass to the node handler.

(And all of this, analgously for the edge file.)

Beside being extremely inefficient (why parse every row twice?), this approach brought other problems with it. Namely, it was error-prone, as the conversion of a KeyedCSVRow to a string did not guarantee that the resulting string was identical with the original string (from which the KeyedCSVRow was created). I had already taken care of a part of this problem some time ago (see commit a36968b52e4684af0db3c859e934b59d635748c5, the thing with the superfuous commas), but this was not enough. The transformation of a KeyedCSVRow back to a string had other problems as well:
- In entries with the character sequence "\n", this sequence was replaced with actual newline characters
- Entries that were enclosed in quotes before, were not enclosed in quotes anymore after the conversion back to a string. This is in particular a problem if the original string, which was enclosed in quotes, contained commas, as such commas would now be interpreted as CSV separators, and therefore the entries mapped to the wrong keys and ArrayIndexOutOfBoundException's would occur.

We could have tried to solve these problems by replacing newlines with "\n" character sequences, inserting quotes around the entries etc., but this is also error-prone and even then there is no guarantee we thought of every possible problem and it would still be inefficient, at any rate.

Thus I chose the "clean" approach to have CSVAST maintain two LinkedList<KeyedCSVRow> (one for the nodes and one for the edges) instead of two LinkedList<String>. It now has functions to return iterators over these KeyedCSVRow's instead of functions that join a LinkedList<String> together with newline characters. All other files that deal with CSVAST's were changed accordingly. CSV2AST's main convert() function now expects a CSVAST, though I implemented methods convert() that take filenames or StringReader's, converting them to CSVAST objects and calling convert() on that. (Before, it was the other way around: CSV2AST's main convert() function operated on StringReader's, and there was a convenience convert() function that took a CSVAST and extracted StringReader's from it.)

Beside being less error-prone and more time-efficient, this design is also more space-efficient, since a KeyedCSVRow needs not maintain a string representation of itself any longer.