Added ClosureExpression class to hold Closure.
Consider:
* A function definition is not an expression;
* a closure is a function definition;
* a closure is also an expression.

Err... what? Maybe we should say "a function definition is not *usually* an expression". Closures are special in that respect.

This is a problem for Joern, as it cannot cast a Closure to an Expression, which would be necessary, e.g., in an assignment of a closure to a variable.

Therefore, analogously to ExpressionStatement which extends Statement and holds an Expression,
I created a ClosureExpression which extends Expression and holds a Closure.

Whenever the edge interpreter encounters an "end node" that is a closure (i.e., it tries to add an edge from some start node, say an AssignmentExpression, to a Closure), it now creates a ClosureExpression holding that Closure to cope with this problem.